+++ JDBC/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=6c8cf851
+build.xml.script.CRC32=0dd7d12b
+build.xml.stylesheet.CRC32=f85dc8f2@1.102.0.48
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=6c8cf851
+nbproject/build-impl.xml.script.CRC32=d7a606d4
+nbproject/build-impl.xml.stylesheet.CRC32=12e0a6c2@1.102.0.48
